Update README.md
# Communications Lab

## Course Description
Communications Lab is a production based course that surveys various technologies including web development, sound, 2D design, digital imaging, video, effects and animation. The forms and uses of these communications technologies are explored in a laboratory context of experimentation, collaboration, and discussion. Much of class time will be spent introducing and playing with equipment and software essential to media production and contemporary storytelling. Each technology is examined as a tool that can be employed and utilized in a variety of situations and experiences. Students will gain fundamental experience thinking, writing, and producing across a variety of media. The world wide web will serve as the primary environment for content delivery and user-interaction. Principles of interpersonal communications and media theory are also introduced with an emphasis on storytelling fundamentals, user-­centered design, and interactivity. Outside of the classroom, students will work both individually and in assigned groups on a variety of assignments. Each major assignment will have a written and practical component. In sum, students will be expected to complete weekly readings, homework exercises, four major group assignments, and one final web portfolio project.

## Learning Outcomes
- Develop a fundamental skill set for producing media assets and writing across a variety of media forms
- Gain experience and confidence working with media production equipment and software
- Learn how to leverage production processes to take an idea and make it come to life
- Learn how to apply storytelling and user­ centered principles towards media production, writing, interaction design, and web design and development
- Learn the fundamentals of client-side web development, specifically HTML, CSS and Javascript
- Gain ability to design and develop web pages that communicate a story by leveraging media and web-based interactions
- Gain experience collaborating with others in both creative and technical processes